[dashboard] be explicit about user ownership
The dashboard-deploy role assumes it's run as the ubuntu user (the ansible user)
and installs the application as the non-privileged user. This is fine, but it's
different than all of our other roles.

Instead of setting `become: no` on the play, assume the role is run with become,
like all of our other roles. When files should be installed as a non-privileged
user, explicitly specify the user/group.